services/ai_service/routers/ocr.py
from fastapi import APIRouter, HTTPException, UploadFile, File
from fastapi.responses import JSONResponse
import logging
from typing import Optional, List
from PIL import Image
import io

logger = logging.getLogger(__name__)
router = APIRouter()


@router.post("/extract-text")
async def extract_text_from_image(
    image_file: UploadFile = File(...), language: Optional[str] = "eng"
):
    """
    Extract text from image using OCR.

    Args:
        image_file: Image file to process
        language: Language code for OCR

    Returns:
        JSON with extracted text and metadata
    """
    try:
        # Validate file type
        if not image_file.content_type.startswith("image/"):
            raise HTTPException(status_code=400, detail="File must be an image file")

        # Read image content
        content = await image_file.read()

        # Validate image format
        try:
            image = Image.open(io.BytesIO(content))
            image.verify()
        except Exception:
            raise HTTPException(status_code=400, detail="Invalid image format")

        # TODO: Implement OCR using PaddleOCR or Tesseract
        # For now, return placeholder response
        extracted_text = "This is placeholder extracted text. Implement OCR here."
        confidence = 0.95

        return JSONResponse(
            {
                "success": True,
                "extracted_text": extracted_text,
                "confidence": confidence,
                "language": language,
                "file_name": image_file.filename,
                "file_size": len(content),
                "image_dimensions": {"width": image.width, "height": image.height},
            }
        )

    except Exception as e:
        logger.error(f"OCR error: {str(e)}")
        raise HTTPException(status_code=500, detail="OCR processing failed")


@router.post("/extract-text-batch")
async def extract_text_from_multiple_images(
    image_files: List[UploadFile] = File(...), language: Optional[str] = "eng"
):
    """
    Extract text from multiple images.

    Args:
        image_files: List of image files to process
        language: Language code for OCR

    Returns:
        JSON with extracted text from all images
    """
    try:
        results = []

        for image_file in image_files:
            # Validate file type
            if not image_file.content_type.startswith("image/"):
                continue

            # Read image content
            content = await image_file.read()

            # TODO: Implement OCR for each image
            extracted_text = f"Placeholder text for {image_file.filename}"
            confidence = 0.95

            results.append(
                {
                    "file_name": image_file.filename,
                    "extracted_text": extracted_text,
                    "confidence": confidence,
                    "file_size": len(content),
                }
            )

        return JSONResponse(
            {
                "success": True,
                "results": results,
                "total_files": len(results),
                "language": language,
            }
        )

    except Exception as e:
        logger.error(f"Batch OCR error: {str(e)}")
        raise HTTPException(status_code=500, detail="Batch OCR processing failed")


@router.get("/supported-languages")
async def get_ocr_supported_languages():
    """
    Get list of supported languages for OCR.

    Returns:
        JSON with supported language codes and names
    """
    languages = [
        {"code": "eng", "name": "English"},
        {"code": "hin", "name": "Hindi"},
        {"code": "tam", "name": "Tamil"},
        {"code": "tel", "name": "Telugu"},
        {"code": "ben", "name": "Bengali"},
        {"code": "mar", "name": "Marathi"},
        {"code": "guj", "name": "Gujarati"},
        {"code": "kan", "name": "Kannada"},
        {"code": "mal", "name": "Malayalam"},
        {"code": "ori", "name": "Oriya"},
        {"code": "pan", "name": "Punjabi"},
        {"code": "urd", "name": "Urdu"},
    ]

    return JSONResponse({"success": True, "languages": languages})


@router.post("/extract-medical-info")
async def extract_medical_information(
    image_file: UploadFile = File(...), language: Optional[str] = "eng"
):
    """
    Extract medical information from documents.

    Args:
        image_file: Medical document image
        language: Language code for OCR

    Returns:
        JSON with extracted medical information
    """
    try:
        # Validate file type
        if not image_file.content_type.startswith("image/"):
            raise HTTPException(status_code=400, detail="File must be an image file")

        # TODO: Implement medical information extraction
        # For now, return placeholder response
        medical_info = {
            "patient_name": "Placeholder Name",
            "device_name": "Placeholder Device",
            "manufacturer": "Placeholder Manufacturer",
            "serial_number": "Placeholder Serial",
            "date": "Placeholder Date",
            "dosage": "Placeholder Dosage",
            "prescription": "Placeholder Prescription",
        }

        return JSONResponse(
            {
                "success": True,
                "medical_information": medical_info,
                "language": language,
                "file_name": image_file.filename,
            }
        )

    except Exception as e:
        logger.error(f"Medical info extraction error: {str(e)}")
        raise HTTPException(
            status_code=500, detail="Medical information extraction failed"
        )
