services/ai_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
gunicorn==23.0.0
whitenoise==6.6.0
Pillow==10.3.0
django-storages==1.14.2
boto3==1.34.69
requests==2.32.4

# Database Dependencies
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0

# ML Dependencies - using pre-built wheels where possible
openai-whisper==20231117
paddleocr==2.7.0.3
transformers==4.48.0
torch==2.2.0+cpu
torchvision==0.17.0+cpu
torchaudio==2.2.0+cpu
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
python-magic==0.4.27
opencv-python-headless==4.6.0.66
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
pytest==7.4.3
